+++ src/api/user/user.controller.ts
@@ -68,28 +68,4 @@ export async function deleteUserHandler(req: Request, res: Response) {
 
   return res.json({ message: 'user deleted' })
 }
-export async function loginHandler(req: Request, res: Response) {
 
-  const { email, password } = req.body;
-
-  try {
-    const user = await getUserByEmail(email)
-    if (!user) {
-      return res.status(404).json({ message: "user not found" })
-    }
-
-    //compare
-    const isMatch = await comparePassword(password, user.password)
-
-    if (!isMatch) {
-      return res.status(400).json({ message: "Incorret email or password" })
-    }
-
-    //jwt
-    return res.json(user)
-
-  } catch (error) {
-
-  }
-
-}
