+++ src/auth/local/index.ts
@@ -1,0 +1,10 @@
+//rutas
+import { Router } from "express";
+import { loginHandler } from "./local.controller";
+const router = Router()
+
+router.post("/login", loginHandler)
+
+//
+//
+export default router
