+++ src/auth/utils/bcrypt.ts
@@ -1,5 +1,5 @@
 import bcrypt from 'bcrypt';
-import crypto from 'crypto';
+//import crypto from 'crypto';
 export async function hashPassword(password: string, factor?: number) {
   const salt = await bcrypt.genSalt()
   return await bcrypt.hash(password, salt)
